pyproject.toml
[project]
name = "recipe-chatbot"
version = "0.1.0"
description = "AI Evaluations Course - Recipe Chatbot with comprehensive evaluation framework"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "litellm[caching]",
    "python-dotenv",
    "httpx",
    "rich",
    "pandas",
    "numpy",
    "scikit-learn",
    "judgy",
    "rank-bm25",
    "tqdm",
    "pydantic",
    "typing-extensions",
    "matplotlib",
    "seaborn",
    "plotly",
    "python-fasthtml",
    "monsterui",
    "langsmith>=0.4.38",
    "jupyter",
    "ipykernel",
    "notebook",
    "claudette>=0.3.6",
    "nltk>=3.8.0",
    "scipy>=1.10.0",
    "trueskill>=0.4.5",
    "python-Levenshtein>=0.21.0",
    "weasyprint>=60.0",
    "langgraph>=0.2.0",
    "openai>=1.0.0",
    "redis>=5.0.0",
    "json-repair>=0.30.0",
    "pinecone-client>=3.0.0",
    "tf-keras>=2.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "isort",
    "flake8",
    "mypy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "nbconvert>=7.0.0",
]

[project.scripts]
recipe-chatbot = "backend.main:app"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.per-file-ignores]
"*.ipynb" = ["E402"]  # Allow imports not at top of file in notebooks (configuration cells come first)
"lesson-14/memory_systems_implementation.ipynb" = ["E402"]  # Notebook has config cells before imports
